--- a/.dockerignore
+++ b/.dockerignore
@@ -1,0 +1,4 @@
+.git
+cache
+var
+capnp-secrets