--- a/Dockerfile
+++ b/Dockerfile
@@ -1,9 +1,7 @@
-FROM ocaml/opam:debian-12-opam AS build
-ENV OPAMYES="1" OPAMCONFIRMLEVEL="unsafe-yes" OPAMERRLOGLEN="0" OPAMPRECISETRACKING="1"
+FROM ocaml/opam:debian-12-ocaml-5.1 AS build
 RUN sudo apt-get update && sudo apt-get install libev-dev capnproto m4 pkg-config libsqlite3-dev libgmp-dev libzstd-dev -y --no-install-recommends
-RUN cd ~/opam-repository && git fetch -q origin master && git reset --hard 7109cb085f1064ae0722bdbdd0edf2d51bbe603b && opam update
+RUN cd ~/opam-repository && git fetch -q origin master && git reset --hard b61304c6db353e679a36720d8b914b029d6fbc0c && opam update
 RUN sudo ln -sf /usr/bin/opam-2.1 /usr/bin/opam
-RUN opam switch create 5.1.0~beta1
 COPY --chown=opam solver-service.opam solver-service-api.opam /src/
 WORKDIR /src
 RUN opam install -y --deps-only .
@@ -12,7 +10,7 @@ RUN opam exec -- dune build @install
 RUN opam exec -- dune install --prefix=/usr/local --destdir=pkg --section=bin --relocatable solver-service
 
 FROM debian:12
-RUN apt-get update && apt-get install libev4 curl git libsqlite3-dev ca-certificates netbase -y --no-install-recommends
+RUN apt-get update && apt-get install libev4 curl git libsqlite3-0 ca-certificates netbase -y --no-install-recommends
 WORKDIR /var/lib/ocluster-worker
 ENTRYPOINT ["/usr/local/bin/solver-service"]
 COPY --from=build \