--- a/service/domain_worker.ml
+++ b/service/domain_worker.ml
@@ -1,3 +1,5 @@
+open Eio.Std
+
 module Worker = Solver_service_api.Worker
 module Solver = Opam_0install.Solver.Make (Git_context)
 
@@ -6,6 +8,7 @@ type request = {
   root_pkgs : (OpamTypes.name * (OpamTypes.version * OpamFile.OPAM.t)) list;
   pinned_pkgs : (OpamTypes.name * (OpamTypes.version * OpamFile.OPAM.t)) list;
   vars : Solver_service_api.Worker.Vars.t;
+  cancelled : unit Eio.Promise.t option;
 }
 
 type reply = ((OpamPackage.t list, string) result * float, [`Msg of string]) result
@@ -26,32 +29,35 @@ let env (vars : Worker.Vars.t) v =
     (* OpamConsole.warning "Unknown variable %S" v; *)
     None
 
-let solve { packages; root_pkgs; pinned_pkgs; vars } =
-  try
-    let pins = root_pkgs @ pinned_pkgs |> OpamPackage.Name.Map.of_list in
-    let root_pkgs = List.map fst root_pkgs in
-    let ocaml_package = OpamPackage.Name.of_string vars.ocaml_package in
-    let ocaml_version = OpamPackage.Version.of_string vars.ocaml_version in
-    let constraints = OpamPackage.Name.Map.singleton ocaml_package (`Eq, ocaml_version) in
-    let with_beta_remote = Ocaml_version.(Releases.is_dev (of_string_exn vars.ocaml_version)) in
-    let context =
-      Git_context.create ()
-        ~packages
-        ~pins
-        ~env:(env vars)
-        ~constraints
-        ~test:(OpamPackage.Name.Set.of_list root_pkgs)
-        ~with_beta_remote
-        ~lower_bound:vars.lower_bound
-    in
-    let t0 = Unix.gettimeofday () in
-    let r = Solver.solve context (ocaml_package :: root_pkgs) in
-    let t1 = Unix.gettimeofday () in
-    let r =
-      match r with
-      | Ok sels -> Ok (Solver.packages_of_result sels)
-      | Error diagnostics -> Error (Solver.diagnostics diagnostics)
-    in
-    Ok (r, (t1 -. t0))
-  with Failure msg ->
-    Error (`Msg msg)
+let solve { packages; root_pkgs; pinned_pkgs; vars; cancelled } =
+  match cancelled with
+  | Some p when Promise.is_resolved p -> Error (`Msg "Cancelled")
+  | _ ->
+    try
+      let pins = root_pkgs @ pinned_pkgs |> OpamPackage.Name.Map.of_list in
+      let root_pkgs = List.map fst root_pkgs in
+      let ocaml_package = OpamPackage.Name.of_string vars.ocaml_package in
+      let ocaml_version = OpamPackage.Version.of_string vars.ocaml_version in
+      let constraints = OpamPackage.Name.Map.singleton ocaml_package (`Eq, ocaml_version) in
+      let with_beta_remote = Ocaml_version.(Releases.is_dev (of_string_exn vars.ocaml_version)) in
+      let context =
+        Git_context.create ()
+          ~packages
+          ~pins
+          ~env:(env vars)
+          ~constraints
+          ~test:(OpamPackage.Name.Set.of_list root_pkgs)
+          ~with_beta_remote
+          ~lower_bound:vars.lower_bound
+      in
+      let t0 = Unix.gettimeofday () in
+      let r = Solver.solve context (ocaml_package :: root_pkgs) in
+      let t1 = Unix.gettimeofday () in
+      let r =
+        match r with
+        | Ok sels -> Ok (Solver.packages_of_result sels)
+        | Error diagnostics -> Error (Solver.diagnostics diagnostics)
+      in
+      Ok (r, (t1 -. t0))
+    with Failure msg ->
+      Error (`Msg msg)