--- a/service/opam_repository.mli
+++ b/service/opam_repository.mli
@@ -1,2 +1,0 @@
-include Opam_repository_intf.S
-(** @inline *)