@mtelvers 2023-07-18
Can the new solver instances be distributed across machines?  Alpha said that the original solver could only have a single instance running servicing the solver pool. With that restriction removed and lower cpu requirements, could we deploy a solver work to many machines in the cluster along side the normal worker?