@moyodiallo 2023-07-18
> Can the new solver instances be distributed across machines? Alpha said that the original solver could only have a single instance running servicing the solver pool. 

@mtelvers are you talking about the pool of process behind the service ? But with the old version it possible to have more than one `solver-worker` connected to the scheduler.

>  could we deploy a solver work to many machines in the cluster along side the normal worker?

Yes we could.