@moyodiallo 2023-07-18
@talex5 Thanks for the port, I read the code it's wonderful. Easier to read than the old version.

There's just one last thing, after that we could deploy it on the prod. It is about the switch when a job is canceled.