@moyodiallo 2023-07-18
> > There's just one last thing, after that we can deploy it on the prod. It is about the switch when a job is canceled.
> 
> Since solve jobs only take about a second to run, I just ignored cancellation. The only thing that takes much time is doing a git-fetch, and it's probably safest not to cancel that anyway.

Yes, in the case of domains there's no way to cancel, we could avoid running a request if its switch is off.